A first end to end implementation.
The notebook that will replace the overview report is prototypes/autoCreateNotebooks.py

To work properly  'jupytext' must be installed (which it should be after "source install_developer_conda.sh" )
Additionally *.ipynb files have to be paied  with the "light script"
version *.py which is the repository version.

The notebook contains some (still dreadfull looking) proof of concept
implementations for interactive elements.

The README now contains some hints for a workflow of translating the
yaml files.

bgc_md2/models/Williams2005GCB/source.py is an example of a real world
model defined in the new format.
The smooth_reservoir_model instance and the model runs are still
missing, but all the variables are defined including some 'specialVars'
for intermodel comparison. Some others are derifed automatically.

tests/TestMVarConstructors.py
Is the nucleus for a more comprehensive examination of the things
possible with this model and probably a the place to check that we
can infer stuff we want to include in the interactive notebook.