added a "1-pixel globe" run
Added a function to model_specific_helpers.R to import data and summarize it as a "1-pixel globe" with total values for C pools and fluxes and mean values for temperature and moisture.
Added some lines to main.R (commented by default) to import data for the globe, and to modify initial values for the globe.