I implemented a variable timestep iterator.
and it works. Even with a 30 day timestep the graphs look very similar.
This has some implications for other parts of the code.

To be able to compare the results for rh  for  iterators with
different timesteps I switched to the computation of the flux
density (Mass/Area/time) instead of an  accumulated density flux
over the timestep (=density = Mass/Area in) the iterator.
(Accumulated results would depend on the lenght of the timestep)
If we really needed an accumulated value (to compare to
observations) we would have to do the accumulation in param2res
I used the netcdf4 package to look at the trendy data for
visit and the reported unit is kg m-2 s-1
This is obviously a flux and not an accumulated value as
we originally assumed (why did we assume this?  )
I commented out the monthly to yearly stuff (since
we technically do not need it, I think you did it)

As a result the iterator as well as param2res become
simpler...but you should have a look at our visit 2
example createModel.py